Use mlx-lm kvcache.
