+++ mlx_audio/tts/models/pocket_tts/codec/mimi/mimi.py
@@ -19,6 +19,14 @@ from .modules import (
 )
 
 
+def _reset_kv_cache(cache) -> None:
+    cache.keys = None
+    cache.values = None
+    cache.offset = 0
+    if hasattr(cache, "_idx"):
+        cache._idx = 0
+
+
 @dataclass
 class MimiConfig:
     channels: int
@@ -127,14 +135,14 @@ class Mimi(nn.Module):
         self.encoder.reset_state()
         self.decoder.reset_state()
         for c in self.decoder_cache:
-            c.reset()
+            _reset_kv_cache(c)
         for c in self.encoder_cache:
-            c.reset()
+            _reset_kv_cache(c)
 
     def encode(self, xs: mx.array) -> mx.array:
         self.encoder.reset_state()
         for c in self.encoder_cache:
-            c.reset()
+            _reset_kv_cache(c)
         xs = self.encoder(xs)
         xs = self.encoder_transformer(xs, cache=self.encoder_cache)[0]
         xs = self.downsample(xs)
@@ -143,7 +151,7 @@ class Mimi(nn.Module):
     def decode(self, xs: mx.array) -> mx.array:
         self.decoder.reset_state()
         for c in self.decoder_cache:
-            c.reset()
+            _reset_kv_cache(c)
         xs = self.quantizer.decode(xs)
         xs = self.upsample(xs)
         xs = self.decoder_transformer(xs, cache=self.decoder_cache)[0]
@@ -194,7 +202,7 @@ class MimiStreamingDecoder:
         self._mimi.decoder.reset_state()
         self._mimi.upsample.reset_state()
         for c in self._mimi.decoder_cache:
-            c.reset()
+            _reset_kv_cache(c)
 
     def decode_frames(self, tokens: mx.array) -> mx.array:
         """Decode a sequence of audio tokens incrementally.
