+++ mlx_audio/tts/models/pocket_tts/codec/mimi/modules/__init__.py
@@ -1,3 +1,5 @@
+from mlx_lm.models.cache import KVCache, RotatingKVCache
+
 from .conv import (
     Conv1d,
     ConvDownsample1d,
@@ -8,7 +10,6 @@ from .conv import (
     StreamableConv1d,
     StreamableConvTranspose1d,
 )
-from .kv_cache import KVCache, RotatingKVCache
 from .quantization import EuclideanCodebook, SplitResidualVectorQuantizer
 from .seanet import SeanetConfig, SeanetDecoder, SeanetEncoder
 from .transformer import ProjectedTransformer, Transformer, TransformerConfig
